split core and adapters
