+++ commit-scoreboard/scoreboard/core/vcs_repos.py
@@ -1,8 +1,4 @@
-import git
-
 from .app import VersionControlRepo
-
-
 
 
 class DummyVersionControlRepo(VersionControlRepo):
@@ -12,11 +8,3 @@ class DummyVersionControlRepo(VersionControlRepo):
     def count_commits_ahead(self, ref: str, target: str) -> int:
         return self.branch_commits[target] - self.branch_commits[ref]
         
-
-class GitVersionControlRepo(VersionControlRepo):
-    def __init__(self, git_repository: git.Repo):
-        self.git_repository = git_repository
-
-    def count_commits_ahead(self, ref: str, target: str) -> int:
-        return len(list(self.git_repository.iter_commits(f"{ref}..{target}")))
-
